Refactor to repository module
