+++ cookiecutter/main.py
@@ -13,58 +13,15 @@ library rather than a script.
 from __future__ import unicode_literals
 import logging
 import os
-import re
 
 from .config import get_user_config, USER_CONFIG_PATH
-from .exceptions import InvalidModeException, RepositoryNotFound
-from .prompt import prompt_for_config
 from .generate import generate_context, generate_files
-from .vcs import clone
+from .exceptions import InvalidModeException
+from .prompt import prompt_for_config
 from .replay import dump, load
+from .repository import determine_repo_dir
 
 logger = logging.getLogger(__name__)
-
-builtin_abbreviations = {
-    'gh': 'https://github.com/{0}.git',
-    'bb': 'https://bitbucket.org/{0}',
-}
-
-REPO_REGEX = re.compile(r"""
-(?x)
-((((git|hg)\+)?(git|ssh|https?):(//)?)  # something like git:// ssh:// etc.
- |                                      # or
- (\w+@[\w\.]+)                          # something like user@...
-)
-""")
-
-
-def is_repo_url(value):
-    """Return True if value is a repository URL."""
-    return bool(REPO_REGEX.match(value))
-
-
-def expand_abbreviations(template, config_dict):
-    """
-    Expand abbreviations in a template name.
-
-    :param template: The project template name.
-    :param config_dict: The user config, which will contain abbreviation
-        definitions.
-    """
-
-    abbreviations = builtin_abbreviations.copy()
-    abbreviations.update(config_dict.get('abbreviations', {}))
-
-    if template in abbreviations:
-        return abbreviations[template]
-
-    # Split on colon. If there is no colon, rest will be empty
-    # and prefix will be the whole template
-    prefix, sep, rest = template.partition(':')
-    if prefix in abbreviations:
-        return abbreviations[prefix].format(rest)
-
-    return template
 
 
 def cookiecutter(
@@ -96,24 +53,13 @@ def cookiecutter(
     # If no config file, sensible defaults from config.DEFAULT_CONFIG are used
     config_dict = get_user_config(config_file=config_file)
 
-    template = expand_abbreviations(template, config_dict)
-
-    if is_repo_url(template):
-        repo_dir = clone(
-            repo_url=template,
-            checkout=checkout,
-            clone_to_dir=config_dict['cookiecutters_dir'],
-            no_input=no_input
-        )
-    else:
-        # If it's a local repo, no need to clone or copy to your
-        # cookiecutters_dir
-        repo_dir = template
-
-    if not os.path.isdir(repo_dir):
-        raise RepositoryNotFound(
-            'The repository {0} could not be located.'.format(template)
-        )
+    repo_dir = determine_repo_dir(
+        template,
+        config_dict.get('abbreviations', {}),
+        config_dict['cookiecutters_dir'],
+        checkout,
+        no_input
+    )
 
     template_name = os.path.basename(os.path.abspath(template))
